refactor: Use BigDecimal for more accurate calculations and change variable type
* Change BigDecimal Type
- LottoEnum Class : value(prize, winnerCount)

* Change Variable Type
- cost : int -> double

+++ src/main/java/lotto/Generate.java
@@ -9,9 +9,9 @@ import java.util.List;
 import static lotto.Constant.*;
 
 public class Generate {
-    int cost;
+    double cost;
 
-    public Generate(int cost) {
+    public Generate(double cost) {
         this.cost = cost;
     }
